Begin reorganizing files; remove unused image
