+++ lib/index.js
@@ -1,3 +1,5 @@
+const Weather = require('./javascript/weather.js').default
+const BackgroundImage = require('./javascript/background-image.js').default
 
 // ------ display signup form -----------
 document.getElementById('signupLink').addEventListener('click', displaySignupForm)
@@ -125,6 +127,3 @@ appendWeather()
 // };
 
 // getWeather();
-
-const Weather = require('./weather.js').default
-const BackgroundImage = require('./background-image.js').default
