+++ lib/javascript/background-image.js
