+++ lib/javascript/weather.js
